refactor: share user provider
