+++ learning-games/src/components/CourseOverview.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react'
 import { Link } from 'react-router-dom'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../../shared/UserContext'
 import COURSES from '../data/courses'
 import Card from './ui/card'
