+++ learning-games/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App'
 import './index.css'
-import { UserProvider } from './context/UserProvider'
+import { UserProvider } from '../../shared/UserProvider'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
