+++ learning-games/src/pages/AgeInputForm.tsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useState } from 'react'
 import type { FormEvent } from 'react'
 import { useNavigate, Link } from 'react-router-dom'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../../shared/UserContext'
 import './AgeInputForm.css'
 
 /**
