+++ learning-games/src/pages/BadgesPage.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react'
 import { Link } from 'react-router-dom'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../../shared/UserContext'
 import { BADGES } from '../data/badges'
 import './BadgesPage.css'
