+++ learning-games/src/pages/PromptDartsGame.tsx
@@ -5,7 +5,7 @@ import ProgressSidebar from '../components/layout/ProgressSidebar'
 import WhyCard from '../components/layout/WhyCard'
 import InstructionBanner from '../components/ui/InstructionBanner'
 import TimerBar from '../components/ui/TimerBar'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../../shared/UserContext'
 import shuffle from '../utils/shuffle'
 import { getTimeLimit } from '../utils/time'
 import { getApiBase } from '../utils/api'
