+++ learning-games/src/pages/QuizGame.tsx
@@ -5,7 +5,7 @@ import { motion } from 'framer-motion'
 import { toast } from 'react-hot-toast'
 import { Link, useNavigate } from 'react-router-dom'
 import CompletionModal from '../components/ui/CompletionModal'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../../shared/UserContext'
 import './QuizGame.css'
 import InstructionBanner from '../components/ui/InstructionBanner'
 import { HALLUCINATION_EXAMPLES } from '../data/hallucinationExamples'
