+++ learning-games/src/pages/SplashPage.tsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useState } from 'react'
 import type { FormEvent } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../../shared/UserContext'
 import './SplashPage.css'
 
 export default function SplashPage() {
