+++ nextjs-app/src/components/layout/ProgressSidebar.tsx
@@ -2,7 +2,7 @@ import { useContext, useEffect, useRef, useState } from 'react'
 import { getApiBase } from '../../utils/api'
 import confetti from 'canvas-confetti'
 import Link from 'next/link'
-import { UserContext } from '../../context/UserContext'
+import { UserContext } from '../../../../shared/UserContext'
 import { getTotalPoints } from '../../utils/user'
 import Tooltip from '../ui/Tooltip'
 import type { PointsEntry } from '../../pages/leaderboard'
