+++ nextjs-app/src/pages/age.tsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useState } from 'react'
 import type { FormEvent } from 'react'
 import Link from 'next/link'; import { useRouter } from 'next/router'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../../shared/UserContext'
 import '../styles/AgeInputForm.css'
 
 /**
