+++ nextjs-app/src/pages/badges.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react'
 import Link from 'next/link'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../../shared/UserContext'
 import { BADGES } from '../data/badges'
 import '../styles/BadgesPage.css'
