+++ nextjs-app/src/pages/games/quiz.tsx
@@ -6,7 +6,7 @@ import { toast } from 'react-hot-toast'
 import Link from 'next/link'; import { useRouter } from 'next/router'
 import CompletionModal from '../../components/ui/CompletionModal'
 import HeadTag from 'next/head'
-import { UserContext } from '../../context/UserContext'
+import { UserContext } from '../../../../shared/UserContext'
 import '../../styles/QuizGame.css'
 import InstructionBanner from '../../components/ui/InstructionBanner'
 import { HALLUCINATION_EXAMPLES } from '../../data/hallucinationExamples'
