+++ nextjs-app/src/pages/games/recipe.tsx
@@ -13,7 +13,7 @@ import InstructionBanner from '../../components/ui/InstructionBanner'
 import Tooltip from '../../components/ui/Tooltip'
 import TimerBar from '../../components/ui/TimerBar'
 import CompletionModal from '../../components/ui/CompletionModal'
-import { UserContext } from '../../context/UserContext'
+import { UserContext } from '../../../../shared/UserContext'
 import { getTimeLimit } from '../../utils/time'
 import '../../styles/PromptRecipeGame.css'
