+++ nextjs-app/src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect } from 'react'
 import HeadTag from 'next/head'
 import Link from 'next/link'; import { useRouter } from 'next/router'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../../shared/UserContext'
 import { getTotalPoints } from '../utils/user'
 import '../styles/Home.css'
 import { GOAL_POINTS } from '../constants/progress'
