+++ nextjs-app/src/pages/leaderboard.tsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useMemo, useState } from 'react'
 import Link from 'next/link'
 import { toast } from 'react-hot-toast'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../../shared/UserContext'
 import ProgressSidebar from '../components/layout/ProgressSidebar'
 import '../styles/LeaderboardPage.css'
 import { getApiBase } from '../utils/api'
