+++ shared/types/user.ts
@@ -1,0 +1,36 @@
+export interface UserData {
+  /** User's name for personalized greetings */
+  name: string | null
+  age: number | null
+  /** Player selected difficulty level */
+  difficulty: 'easy' | 'medium' | 'hard'
+  points: Record<string, number>
+  badges: string[]
+}
+
+export interface UserContextType {
+  user: UserData
+  setUser: (user: UserData) => void
+  /**
+   * Update only the age field of the user. Games read this to tailor
+   * content difficulty and themes.
+   */
+  setAge: (age: number) => void
+  /**
+   * Update only the name field of the user.
+   */
+  setName: (name: string) => void
+  /**
+   * Record points for a specific game. Implementations typically
+   * store the best/highest points seen so far.
+   */
+  setPoints: (game: string, points: number) => void
+  /**
+   * Award a badge when the player reaches a milestone.
+   */
+  addBadge: (badge: string) => void
+  /**
+   * Update the difficulty setting for all games.
+   */
+  setDifficulty: (level: 'easy' | 'medium' | 'hard') => void
+}
